chore: undo breakup of duration.ts file into 2 parts, to make minimal changes for code reviewers
