+++ src/shared/utils/duration.test.tsx
@@ -5,7 +5,7 @@ import {
   millisecondsToDuration,
   isDurationWithNowParseable,
   isDurationParseable,
-} from 'src/shared/utils/duration/utils'
+} from 'src/shared/utils/duration'
 import {SELECTABLE_TIME_RANGES} from 'src/shared/constants/timeRanges'
 
 import {CustomTimeRange, SelectableDurationTimeRange} from 'src/types'
