--- src/shared/utils/duration/index.ts
@@ -1,6 +1,0 @@
-// legacy code
-// is a blend of timeRange and duration
-// was all in a single file `src/shared/utils/duration`
-
-export * from './utils'
-export * from './timeRange'
